Changed the main menu modal to a screen
